Fixes for few issues handling Koski-integration
